Update : Feature #8 => Android UI 안정화 및 최적화
